[Chore #154082903] Refactor reducers by splitting each reducer into its own file.
